+++ projects/creature-construction-lab/Model/CreatureFeatureType.cs
@@ -3,5 +3,6 @@ namespace CreatureConstructionLab.Model;
 public enum CreatureFeatureType
 {
     Eye,
-    ForkedTongue
+    ForkedTongue,
+    Fin
 }